Secrue API Endpoints, closes #61

+++ src/main/java/ch/zhaw/rentmybike/controller/MotorcycleController.java
@@ -17,6 +17,7 @@ import ch.zhaw.rentmybike.model.dtos.CreateMotorcycleDTO;
 import ch.zhaw.rentmybike.model.entities.Motorcycle;
 import ch.zhaw.rentmybike.repository.MotorcycleRepository;
 import ch.zhaw.rentmybike.services.MotorcycleService;
+import ch.zhaw.rentmybike.services.RoleService;
 
 @RestController
 @RequestMapping("/api/motorcycles")
@@ -28,21 +29,37 @@ public class MotorcycleController {
     @Autowired
     private MotorcycleService motorcycleService;
 
+    @Autowired
+    private RoleService roleService;
+
     // Erstellen eines neuen Motorrads
     @PostMapping("/create")
     public ResponseEntity<Motorcycle> createMotorcycle(@RequestBody CreateMotorcycleDTO motorcycleDTO) {
+
+        if (!(roleService.userHasRole("user") )) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         Motorcycle savedMotorcycle = motorcycleService.createMotorcycle(motorcycleDTO);
         return ResponseEntity.status(HttpStatus.CREATED).body(savedMotorcycle);
     }
     // Abrufen aller Motorräder
     @GetMapping
     public ResponseEntity<List<Motorcycle>> getAllMotorcycles() {
+
+        if (!(roleService.userHasRole("admin"))) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         List<Motorcycle> motorcycles = motorcycleRepository.findAll();
         return ResponseEntity.ok(motorcycles);
     }
     // Abrufen eines Motorrads basierend auf der ID
     @GetMapping("/{id}")
     public ResponseEntity<Motorcycle> getMotorcycleById(@PathVariable String id) {
+
+        if (!(roleService.userHasRole("user") || roleService.userHasRole("admin"))) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         Optional<Motorcycle> motorcycle = motorcycleRepository.findById(id);
         return motorcycle.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
@@ -50,6 +67,11 @@ public class MotorcycleController {
      // Abrufen aller Motorräder eines bestimmten Nutzers
     @GetMapping("/user/{userId}")
     public ResponseEntity<List<Motorcycle>> getMotorcyclesByUserId(@PathVariable String userId) {
+
+
+        if (!(roleService.userHasRole("admin"))) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         List<Motorcycle> motorcycles = motorcycleRepository.findByUserId(userId);
         return ResponseEntity.ok(motorcycles);
     }
@@ -58,6 +80,11 @@ public class MotorcycleController {
     // Löschen eines Motorrads
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteMotorcycle(@PathVariable String id) {
+
+        if (!(roleService.userHasRole("user") || roleService.userHasRole("admin"))) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         Optional<String> motorcycle = motorcycleService.deleteMotorcycleById(id);
         if (motorcycle.isPresent()) {
             motorcycleRepository.deleteById(id);
