+++ src/main/java/ch/zhaw/rentmybike/controller/UserController.java
@@ -2,8 +2,6 @@ package ch.zhaw.rentmybike.controller;
 
 import java.util.Optional;
 
-
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -22,24 +20,30 @@ import ch.zhaw.rentmybike.model.dtos.CreateUserDTO;
 import ch.zhaw.rentmybike.model.entities.User;
 import ch.zhaw.rentmybike.model.entities.User.UserState;
 import ch.zhaw.rentmybike.repository.UserRepository;
+import ch.zhaw.rentmybike.services.RoleService;
 import ch.zhaw.rentmybike.services.UserService;
 
 @RestController
 @RequestMapping("api/users")
 public class UserController {
 
-
     @Autowired
     private UserRepository userRepository;
     @Autowired
     private UserService userService;
+    @Autowired
+    private RoleService roleService;
 
     // alle User ausgeben
     @GetMapping("/all")
     public ResponseEntity<Page<User>> getAllUsers(
-        @RequestParam(required = false) UserState state,
-        @RequestParam(required = false, defaultValue = "1") Integer pageNumber,
-        @RequestParam(required = false, defaultValue = "9") Integer pageSize){
+            @RequestParam(required = false) UserState state,
+            @RequestParam(required = false, defaultValue = "1") Integer pageNumber,
+            @RequestParam(required = false, defaultValue = "9") Integer pageSize) {
+
+        if (!(roleService.userHasRole("admin"))) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         Page<User> allUsers;
         if (state == null) {
             allUsers = userRepository.findAll(PageRequest.of(pageNumber - 1, pageSize));
@@ -52,6 +56,10 @@ public class UserController {
     // User-Daten anhand der ID ausgeben
     @GetMapping("/{userId}")
     public ResponseEntity<User> getUserById(@PathVariable String userId) {
+
+        if (!(roleService.userHasRole("user") || roleService.userHasRole("admin"))) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         Optional<User> optUser = userRepository.findById(userId);
         if (optUser.isPresent()) {
             return new ResponseEntity<>(optUser.get(), HttpStatus.OK);
@@ -70,6 +78,10 @@ public class UserController {
     // User anhand der ID löschen
     @DeleteMapping("/{userId}")
     public ResponseEntity<Void> deleteUserById(@PathVariable String userId) {
+
+        if (!(roleService.userHasRole("admin"))) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         Optional<User> optUser = userRepository.findById(userId);
         if (optUser.isPresent()) {
             userRepository.deleteById(userId);
@@ -80,4 +92,3 @@ public class UserController {
     }
 
 }
-
